fix: move tickets cache to a separate file
